Fixed switch colors and put lab dashboard in place
